pyproject.toml
[project]
name = "drf-example"
version = "0.1.0"
description = "django example"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt>=5.0.0",
    "django>=6.0.1",
    "djangorestframework>=3.16.1",
    "drf-yasg>=1.21.14",
    "jwt>=1.4.0",
    "passlib>=1.7.4",
    "pytest-asyncio>=1.3.0",
    "pytest-django>=4.11.1",
    "python-dotenv>=1.2.1",
    "pytz>=2025.2",
]

[dependency-groups]
dev = ["ruff"]


[tool.ruff]
line-length = 79
indent-width = 4
exclude = [
    "alembic",
    "migration",
    "migrations"
]


[tool.ruff.lint]
extend-select = ["I", "F"]
fixable = ["I"]

[tool.ruff.lint.isort]
combine-as-imports = true

[tool.ruff.lint.pycodestyle]
max-doc-length = 79

